+++ app/__init__.py
@@ -1,6 +1,12 @@
 from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
 
 from config import config
+
+
+db = SQLAlchemy()
+migrate = Migrate()
 
 
 def create_app(config_name):
@@ -12,9 +18,15 @@ def create_app(config_name):
     app.debug = app.config['DEBUG']
     app.secret_key = app.config['SECRET_KEY']
 
+    db.init_app(app)
+    migrate.init_app(app, db)
+
     with app.app_context():  # This makes config object available
         from .main import \
             main as main_blueprint  # These imports need to stay at the end of the function to prevent circular imports
         app.register_blueprint(main_blueprint)
+
+        from .auth import auth as auth_blueprint
+        app.register_blueprint(auth_blueprint, url_prefix='/auth')
 
     return app