+++ config.py
@@ -1,6 +1,9 @@
 class Config(object):
     SERVER_ADDRESS = '127.0.0.1:5000'
     DEBUG = True
+    SQLALCHEMY_DATABASE_URI = 'postgresql://postgres:root@127.0.0.1/travel_arrangements'
+    SQLALCHEMY_TRACK_MODIFICATIONS = True
+    SECRET_KEY = 'd9030f835a1644cfa30e8f832b436c47'
 
     @staticmethod
     def init_app():
@@ -8,11 +11,12 @@ class Config(object):
 
 
 class DevelopmentConfig(Config):
-    SECRET_KEY = 'd9030f835a1644cfa30e8f832b436c47'
+    pass
 
 
 class TestingConfig(Config):
-    pass
+    SQLALCHEMY_DATABASE_URI = 'postgresql://postgres:root@127.0.0.1/test_travel_arrangements'
+    PRESERVE_CONTEXT_ON_EXCEPTION = False
 
 
 class ProductionConfig(Config):